+++ src/main/java/com/deque/axe/ResultType.java
@@ -1,8 +1,4 @@
-/*
- * Copyright 2020 (C) Magenic, All rights Reserved
- */
-
-package com.magenic.jmaqs.accessibility;
+package com.deque.axe;
 
 /**
  * Result Type placement holder.
