+++ src/main/java/com/deque/axe/providers/EmbeddedResourceAxeProvider.java
@@ -1,8 +1,4 @@
-/*
- * Copyright 2020 (C) Magenic, All rights Reserved
- */
-
-package com.magenic.jmaqs.accessibility.providers;
+package com.deque.axe.providers;
 
 import java.io.IOException;
 
@@ -14,6 +10,6 @@ public class EmbeddedResourceAxeProvider implements IAxeScriptProvider {
    * Reads the axe.min.js finder.
    */
   public String getScript() throws IOException {
-    return EmbeddedResourceProvider.readEmbeddedFile("src/test/resources/js/axe.min.js");
+    return EmbeddedResourceProvider.readEmbeddedFile("src/test/resources/files/axe.min.js");
   }
 }
