+++ src/main/java/com/deque/axe/providers/IAxeScriptProvider.java
@@ -1,8 +1,4 @@
-/*
- * Copyright 2020 (C) Magenic, All rights Reserved
- */
-
-package com.magenic.jmaqs.accessibility.providers;
+package com.deque.axe.providers;
 
 import java.io.IOException;
 import javax.naming.OperationNotSupportedException;
