+++ src/test/java/com/deque/axe/AxeExampleTest.java
@@ -1,0 +1,4 @@
+package com.deque.axe;
+
+public class AxeExampleTest {
+}
