+++ src/test/resources/files/Integration-test-target.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>Test page with axe-core violations for integration test</title>
+</head>
+<body>
+<main>
+    <ul>
+        <li>Example violation 1: <span role="invalid">invalid role attribute</span></li>
+        <li>Example violation 2: <span style="color: #aaa">low color contrast</span></li>
+        <li>Example violation 3: <button aria-hidden="true">aria-hidden element is focusable</button></li>
+    </ul>
+</main>
+</body>
+</html>
