--- src/test/resources/timeout.js
@@ -1,8 +1,0 @@
-// A fake axe-core that takes 5 minutes to finish
-window.axe = {
-  run: function (context, options, callback) {
-    setTimeout(function () {
-      callback()
-    }, 1000 * 60 * 5)
-  }
-}
